Fix DTOS

+++ api/dev-test/src/main/java/com/vulcan/dev_test/domain/student/rest/request/StudentDto.java
@@ -2,10 +2,8 @@ package com.vulcan.dev_test.domain.student.rest.request;
 
 import com.vulcan.dev_test.domain.student.Gender;
 
-import java.math.BigInteger;
-
 public record StudentDto(
-        BigInteger id,
+        Long id,
         String firstName,
         String lastName,
         Integer age,
